Fixed more int64->int32 implicit warnings, both in Eternity and in libraries
Except zlib, which is no longer used by the xcodeproj.

Also hopefully skipped the recommended xcode settings to restore the undesired comma warning.

+++ source/textscreen/txt_sdl.c
@@ -954,7 +954,7 @@ int TXT_vsnprintf(char *buf, size_t buf_len, const char *s, va_list args)
     if (result < 0 || result >= buf_len)
     {
         buf[buf_len - 1] = '\0';
-        result = buf_len - 1;
+        result = (int)(buf_len - 1);
     }
 
     return result;
